Add timeout, pause even in failure
Adds a timeout to all home assistant requests.  I think this is hanging
sometimes on networking hiccups.  requests doesn't have a timeout by
default (and no way to blanket apply a timeout to a Session), so add a
wrapper for Session that adds the timeout.

Also moves the 'sleep' in the main loop outside the try/except, so will
always pause between requests, even in the case of failure.